Corrected the data type for "year5", "year7", and "year10"

+++ Ocean_Levels_Project/main.cpp
@@ -5,9 +5,9 @@ using namespace std;
 
 int main() {
 	double lvlC = 1.5; // this is the distance (in millimeters) the ocean rises in a year
-	double year5 = lvlC * 5;
-	double year7 = lvlC * 7;
-	double year10 = lvlC * 10;
+	int year5 = lvlC * 5;
+	int year7 = lvlC * 7;
+	int year10 = lvlC * 10;
 	cout << "In five years, the ocean will have risen " << year5 << " millimeters.\n";
 	cout << "In seven years, the ocean will have risen " << year7 << " millimeters.\n";
 	cout << "In ten years, the ocean will have risen " << year10 << " millimeters.\n";
